% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcsvd.R
\name{ggcsvd}
\alias{ggcsvd}
\title{Groups sparse Generalized SVD of a matrix}
\usage{
ggcsvd(X, R = 2, au = rep(sqrt(nrow(X)), R), av = rep(sqrt(ncol(X)),
  R), Wu = rep(1, NROW(X)), Wv = rep(1, NCOL(X)), Gu = seq(NROW(X)),
  Gv = seq(NCOL(X)), itermax.pi = 1000, itermax.pocs = 1000,
  eps.pi = 1e-16, eps.pocs = 1e-16, init = "svd", order_sv = TRUE)
}
\arguments{
\item{X}{a (data) matrix;}

\item{R}{the desired rank of the singular decomposition;}

\item{au}{The radiuses (>0) of the
$L_1$ ball for each left vector.
Default to the maximum possible radius, such that the result
is the same as the result of a regular SVD.}

\item{av}{The radiuses
(>0) of the $L_1$ balls for each right vector.
Default to the maximum possible radius, such that the result
is the same as the result of a regular SVD.}

\item{Wu}{a vector containing the weights for the lines.}

\item{Wv}{a vector containing the weights for the columns.}

\item{Gu}{a vector describing the groups for the lines.}

\item{Gv}{a vector describing the groups for the columns.}

\item{itermax.pi}{The maximum number of iterations for the power iteration.}

\item{itermax.pocs}{The maximum number of iterations for POCS.}

\item{eps.pi}{Precision for the power iteration.}

\item{eps.pocs}{Precision for POCS.}

\item{init}{How to initialize the algorithm.
Either "svd" (default) or "rand" to intialize with,
respectively, the results of a regular SVD or random
vectors.}

\item{order_sv}{Boolean. Should the singular values
be artificially ordered. Default to TRUE.}
}
\value{
Pseudo-singular vectors (U and V) and values (D), and the number of iterations.
}
\description{
Groups sparse Generalized SVD of a matrix
}
\examples{
X <- matrix(rnorm(20), 5, 4)
ggcsvd(X) # By default, no sparsity constraint is imposed on the decomposition
svd(X) # Should give roughly the same results
}
\author{
Vincent Guillemot
}
