% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powit1.R
\name{powit1}
\alias{powit1}
\title{Power iteration engine, without orthogonality constraints}
\usage{
powit1(X, U0, V0, au, av, Gu, Gv, eps.pi, itermax.pi)
}
\arguments{
\item{itermax.pi}{}
}
\description{
Power iteration engine, without orthogonality constraints
}
