% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powit2.R
\name{powit2}
\alias{powit2}
\title{Power iteration engine, with orthogonality constraints}
\usage{
powit2(X, U0, V0, Uorth, Vorth, au, av, Gu, Gv, eps.pi, eps.pocs,
  itermax.pi, itermax.pocs)
}
\arguments{
\item{itermax.pocs}{}
}
\description{
Power iteration engine, with orthogonality constraints
}
