% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projgrouporth.R
\name{projgrouporth}
\alias{projgrouporth}
\title{Compute the projection of a vector onto the intersection of the
group l1-l2 ball, the l2 ball and the space orthogonal to M.}
\usage{
projgrouporth(x = x, r = 1, g = g, M = M, itermax = 5000,
  eps = 1e-16)
}
\arguments{
\item{x}{A vector of numerics}

\item{r}{The radius (>0) of the l1 ball}

\item{g}{Vector of groups.}

\item{M}{A matrix of vectors}

\item{itermax}{The maximum number of iterations}

\item{eps}{Precision}
}
\value{
The projection of \eqn{x} onto \eqn{B_1(a) \cap B_2 \cap M^\perp}.
}
\description{
Compute the projection of a vector onto the intersection of the
group l1-l2 ball, the l2 ball and the space orthogonal to M.
}
\examples{
projgrouporth(x=1:10, r=10, g = 1:10, M=projl2(rnorm(10)))
}
