% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projl2.R
\name{projl2}
\alias{projl2}
\title{Normalize a vector.}
\usage{
projl2(u)
}
\arguments{
\item{u}{A vector of numerics}
}
\value{
The normalized version of \eqn{u}.
}
\description{
Normalize a vector.
}
\examples{
normalize(c(0, 0,-0.5, 0.5))
}
