% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxgroup.R
\name{proxgroup}
\alias{proxgroup}
\title{Proximal operator of the group l1l2 norm}
\usage{
proxgroup(x, g, lambda)
}
\arguments{
\item{x}{vector of numerics}

\item{g}{vector describing the non-overlapping groups}

\item{lambda}{regularization factor}
}
\value{
The transformation of the input vector by the proximal operator of the group l1l2 norm
}
\description{
Proximal operator of the group l1l2 norm
}
\examples{
x <- rnorm(50)
g <- gl(10, 5)
y <- proxgroup(x, g, 2)
layout(t(1:2))
plot(x, col = g)
plot(y, col = g)

}
