% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxl1.R
\name{proxl1}
\alias{proxl1}
\title{Proximal operator of the l1 norm}
\usage{
proxl1(x, lambda)
}
\arguments{
\item{x}{a vetor of numeric values}

\item{lambda}{}
}
\description{
Proximal operator of the l1 norm
}
