% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxl2.R
\name{proxl2}
\alias{proxl2}
\title{Proximal operator of the l2 norm}
\usage{
proxl2(x, lambda)
}
\arguments{
\item{x}{a vetor of numeric values}

\item{lambda}{}
}
\description{
Proximal operator of the l2 norm
}
