% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssq.R
\name{ssq}
\alias{ssq}
\title{Compute the sum of squares.}
\usage{
ssq(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
The sum of the squared coefficients of vector x.
}
\description{
Compute the sum of squares.
}
\examples{
ssq(1:10)
}
