% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projgroup.R
\name{projgroup}
\alias{projgroup}
\title{Compute the projection of a vector onto the l1 ball.}
\usage{
projgroup(v, a, g)
}
\arguments{
\item{v}{A vector of numerics}

\item{a}{The radius (>0) of the l1 ball}

\item{g}{A vector of groups}
}
\value{
The projection of \eqn{v} onto the l1 ball of radius \eqn{a}.
}
\description{
Compute the projection of a vector onto the l1 ball.
}
\examples{
projgroup(1:10, rep(1:2, e=5), 3)
}
